+++ packages/genesys/packages/react-components-core/src/Hash/index.ts
@@ -1,0 +1,2 @@
+export * from "./Hash.types";
+export type { HashEllipsis } from "@peersyst/react-utils";