+++ packages/genesys/packages/react-components/src/BlockchainAddress/BlockchainAddress.tsx
@@ -4,23 +4,20 @@ import { cx } from "@peersyst/react-utils";
 import { useComponentConfig, useMergeDefaultProps } from "@peersyst/react-components-core";
 
 const BlockchainAddress = (props: BlockchainAddressProps): JSX.Element => {
-    const {
-        address,
-        type,
-        className,
-        copy,
-        link = true,
-        ...hashProps
-    } = useMergeDefaultProps("BlockchainAddress", props);
+    const { address, type, className, addressToShareData, ...hashProps } = useMergeDefaultProps(
+        "BlockchainAddress",
+        props,
+    );
 
     const { blockchainLinks } = useComponentConfig("BlockchainAddress");
+    const url = blockchainLinks[type] + address;
 
     return (
         <Hash
             className={cx("BlockchainAddress", className)}
+            hashToShareData={addressToShareData}
             hash={address}
-            url={link ? blockchainLinks[type] + address : undefined}
-            copy={copy}
+            url={url}
             {...hashProps}
         />
     );