+++ packages/genesys/packages/react-components/src/Hash/Hash.styles.ts
@@ -1,5 +1,5 @@
 import styled, { css } from "styled-components";
-import { HashRootProps, HashTextProps, HashLinkProps } from "./Hash.types";
+import { HashRootProps, HashTextProps } from "./Hash.types";
 import { Row } from "../Row";
 import { Typography } from "../Typography";
 
@@ -16,15 +16,15 @@ export const HashRoot = styled(Row).attrs({
     `,
 );
 
-export const HashLink = styled.a<HashLinkProps>(
-    ({ url }) => css`
-        text-decoration: none;
-        &:hover {
-            text-decoration: ${url ? "underline" : "none"};
-        }
-    `,
-);
-
 export const HashText = styled(Typography)<HashTextProps>`
     word-break: ${(p) => p.break && "break-word"};
 `;
+
+export const HashLink = styled.a(
+    () => css`
+        text-decoration: none;
+        &:hover {
+            text-decoration: underline;
+        }
+    `,
+);