+++ packages/genesys/packages/react-native-components/src/config/components.config.tsx
@@ -290,6 +290,9 @@ const componentsConfig: ComponentsConfig = {
     ContainedSuspense: {
         defaultProps: {},
     },
+    Hash: {
+        defaultProps: {},
+    },
 };
 
 export default componentsConfig;