+++ packages/genesys/packages/react-native-components/src/util/CopyButton/CopyButton.tsx
@@ -1,22 +1,20 @@
-import * as Clipboard from "expo-clipboard";
 import { CopyButtonProps } from "./CopyButton.types";
-import { useToast } from "../../feedback/ToastProvider";
 import { IconButton } from "../../input/IconButton";
 import { CopyIcon } from "../../assets/icons";
 import { useMergeDefaultProps } from "@peersyst/react-components-core";
+import useCopyToClipboard from "../../hooks/useCopyToClipboard/useCopyToClipboard";
 
 const CopyButton = (props: CopyButtonProps): JSX.Element => {
     const { text, style, message, ...rest } = useMergeDefaultProps("CopyButton", props);
 
-    const { showToast } = useToast();
-
-    const copyToClipboard = () => {
-        Clipboard.setStringAsync(text);
-        if (message) showToast(message, { type: "success" });
-    };
+    const copyToClipboard = useCopyToClipboard();
 
     return (
-        <IconButton {...rest} style={style} onPress={() => copyToClipboard()}>
+        <IconButton
+            {...rest}
+            style={style}
+            onPress={() => copyToClipboard({ toastMessage: message, text })}
+        >
             <CopyIcon />
         </IconButton>
     );