+++ packages/genesys/packages/react-native-utils/src/index.ts
@@ -1,2 +1,3 @@
 export { default as extractTextStyles } from "./extractTextStyles";
 export { default as formatNumber } from "./formatNumber";
+export { default as Clipboard } from "./Clipboard";